Added paragraph about how I remember Marshall fondly.
This is the paragraph:

==A time of spiritual growth==

The time I was with Marshall was a very key part of my spiritual
growth. We would have long discussions, sometimes arguments, about
God. When I started “working” for Marshall; I was an atheist. Just
after I broke contact with him, I had a conversion experience and became
a Christian; that experience was how I went from being a child to being
an adult. This experience would not had been possible had Marshall not
been a part of my life.

It is a time of my life I remember very fondly.