Bernie Sanders picture: Add attribution
Also, I am about 99% done making the *entire* blog fit on a phone
screen. There are still some entries from 2020, 2012, and 2011 which
are ever slightly wider than the phone screen, but all of the really
wide elements have been altered to fit on a phone screen.

The last 1% isn’t worth it for the time being; the entire blog is
now phone compatible, quite a feat considering it was written for
years w/o phones in mind.