lunacyBlogFilter: Bug fixes, hotlink links in footnotes
Note: It’s a bad idea to have more than one footnote in a given
paragraph; the footnote processing code is buddy. No HTML tags can
be in footnotes but bare links are converted in to hotlinks.

I also fixed some bugs with spacing around links and try to have the end
of a link end a line in the processed HTML output.