lunacyBlogFilter: Update code to break long words on phone screens
This way, we can have
supercalifragilisticexpialidociousantidisestablishmentarianism
in blog entries and the code will put an optional invisible split the
word so it doesn’t cause layout issues on phones.

The code actually doesn’t break real words; even
“supercalifragilisticexpialidocious” and
“antidisestablishmentarianism” are short enough to not have a break
added to them. This is mainly for things like long bare URLs and what
not.