lunacyBlogFilter: Our regexes can now handle the non-ASCII correctly
This is because we convert the entities in to a special 7-bit format,
so that Lunacy’s 8-bit regex engine can process it, then we convert
it in to UTF-8 when outputting it.

Here is how the non-ASCII is encoded:

  local h={}
  --[[ NUL ]]    h["Á"] = '\1'  h["É"] = '\2'  h["Í"] = '\3'
  h["Ñ"] = '\4'  h["Ó"] = '\5'  h["Ú"] = '\6'  h["Ü"] = '\7'
  h["¡"] = '\8'  --[[ \t ]]     --[[ \n ]]     h["—"] = '\11'
  --[[ \f ]]     --[[ \r ]]     h["«"] = '\14' h["»"] = '\15'
  h["•"] = '\16' h["á"] = '\17' h["é"] = '\18' h["í"] = '\19'
  h["ñ"] = '\20' h["ó"] = '\21' h["ú"] = '\22' h["ü"] = '\23'
  h["¿"] = '\24' h["‘"] = '\25' h["’"] = '\26' --[[ ESC ]]
  h["→"] = '\28' h["“"] = '\29' h["”"] = '\30' h["©"] = '\31'
  h["♥"] = '\127'

(Note that “—” — the emdash — is considered a whitespace with the “%s”
 character class. The workaround is [\t\r\n\f ] for pure whitespace.
Since — does split words, it doesn’t matter that much. Also, characters
inside HTML tages are *not* converted in to this “Xascii” format.