htmldoc really doesn’t like UTF-8
So, as a workaround, we convert the UTF-8 in to a slight variant
of Xascii (where — is \128 instead of \11 because htmldoc considers
\11 a white space character), have htmldoc process it (where it will
leave the control and hi bit characters as is), then process it.

lunacyBlogFilter converts the Xascii back in to UTF-8.  Since
lunacyBlogFilter *already* uses Xascii to process the file internally, it
handles Xascii input just fine; we only needed to add three lines of
code which converts \128 in to the \11 lunacyBlogFilter uses.

This is *ugly* but it works.