numtoX3 This is a “fantasy” format for storing natural numbers
Like UTF-8, this allows ASCII text to be an array of natual numbers.
Unlike UTF-8, the numbers can be arbitrarily large, and there is only
one way to represent a given number. Also, unlike UTF-8, each byte
gives us six (not five) bits of data while making it easy to find the
end of a byte sequence representing a number.

Like UTF-8 this format is fully ASCII compatible.