Add footnote about the subreddits in question
All articles made the front page of Reddit.

As an aside, the reason for the — before the footnote is because of a
combination of bugs make this the way I can put a footnote right after
a link. I will look in to a better way to fix this bug, but this
workaround works for now.