lunacyBlogFilter: Allow multiple footnotes together
We still have a limit of two footnotes on a single line, but with
great effort, we can now have 1,2 style footnotes.  There are some
limitations:

* We can only have 9 footnotes in a blog entry, labeled 1,2,3,4,etc.
  (If we *must* have more, the entry after footnote #9 would be “a”
  with the current codebase, not 10 as is standard)
* We can only have two footnotes together like that

For simple 500 word blog entries, this is good enough for now.

To fix the 9 footnote limit, we would write a sort function to sort on
the numerical value of a string when calling sPairs when grabbing
footnotes.

To fix the two footnotes per line limit (i.e. max two footnotes together),
we would need to rewrite the footnote processing code.