Add graph comparing 18-29 year olds in the NSFG 2022 data
As we can see, the agreement between the number of men with a lot of
sex partners compared to the number of women with a lot of sex partners
is about the same.

The difference between the 18-29 segment and all ages is that the
promiscuous people are more promiscuous.  Indeed, there are more virgins
(100 minus the point on the right hand side is the number of virgins)
with all people than with younger people.